+++ lib/src/hooks/use_form.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/widgets.dart';
-import 'package:flutter_hook_form/src/models/field_schema.dart';
 import 'package:flutter_hook_form/src/models/form_field_controller.dart';
+import 'package:flutter_hook_form/src/validators/field_config.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 
 /// A hook that provides a [FormFieldsController] to manage form field states.
@@ -12,14 +12,18 @@ import 'package:flutter_hooks/flutter_hooks.dart';
 /// **Careful !**: this hook is a `flutter_hooks` hook and needs to be used
 /// inside a [HookWidget]. For more information about `flutter_hooks`, please
 /// refer to the [flutter_hooks documentation](https://pub.dev/packages/flutter_hooks).
-FormFieldsController<F> useForm<F extends FieldSchema<dynamic>>({
-  InitialFieldValues<F, dynamic>? initialValues,
+FormFieldsController<E> useForm<E extends Enum>({
+  Map<E, FieldConfig> validators = const {},
   List<Object?> keys = const <Object>[],
+  bool focusOnInvalid = false,
+  bool autoScrollWhenFocusOnInvalid = true,
 }) {
   final form = useMemoized(() {
-    return FormFieldsController<F>(
+    return FormFieldsController<E>(
       GlobalKey<FormState>(debugLabel: 'FormFieldsController'),
-      initialValues: initialValues,
+      validators: validators,
+      focusOnInvalid: focusOnInvalid,
+      autoScrollWhenFocusOnInvalid: autoScrollWhenFocusOnInvalid,
     );
   }, keys);