+++ lib/src/validators/cross_field_validators.dart
@@ -1,12 +1,16 @@
+// ignore_for_file: public_member_api_docs
+
 import 'package:flutter_hook_form/flutter_hook_form.dart';
 
 /// Compare the value with the value of the cross field and validate if the
 /// value is after the cross field value.
-class DateAfterValidator extends CrossFieldValidator<DateTime> {
-  /// Creates a [IsAfterValidator].
-  const DateAfterValidator({required super.field, super.message})
-    : super(errorCode: 'date_after');
-
+class DateAfterValidator<E extends Enum>
+    extends CrossFieldValidator<DateTime, E> {
+  const new({
+    required super.field,
+    super.message,
+    String? errorCode,
+  }) : super(errorCode: errorCode ?? 'date_after');
   @override
   CrossFieldValidatorFn<DateTime> get validator {
     return (value, context) {
@@ -16,8 +20,8 @@ class DateAfterValidator extends CrossFieldValidator<DateTime> {
 
       assertValueIsOfType(context);
 
-      final form = useFormContext(context);
-      final fieldValue = form.getValue<DateTime>(field as FieldSchema<DateTime>);
+      final form = useFormContext<E>(context);
+      final fieldValue = form.getValue<DateTime>(field);
 
       if (fieldValue != null && value.isBefore(fieldValue)) {
         return message ?? errorCode;
@@ -30,18 +34,17 @@ class DateAfterValidator extends CrossFieldValidator<DateTime> {
 
 /// Compare the value with the value of the cross field and validate if the
 /// value matches the cross field value.
-class MatchesValidator<T> extends CrossFieldValidator<T> {
-  /// Creates a [MatchesValidator].
-  const MatchesValidator({required super.field, super.message})
-    : super(errorCode: 'field_does_not_match');
+class MatchesValidator<T, E extends Enum> extends CrossFieldValidator<T, E> {
+  const new({required super.field, super.message, String? errorCode})
+    : super(errorCode: errorCode ?? 'field_does_not_match');
 
   @override
   CrossFieldValidatorFn<T> get validator {
     return (value, context) {
       assertValueIsOfType(context);
 
-      final form = useFormContext(context);
-      final fieldValue = form.getValue<T>(field as FieldSchema<T>);
+      final form = useFormContext<E>(context);
+      final fieldValue = form.getValue<T>(field);
 
       if (fieldValue != value) {
         return message ?? errorCode;