+++ lib/src/widget/hooked_form.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_hook_form/src/models/field_schema.dart';
 
 import '../hooks/use_form_context.dart';
 import '../models/form_field_controller.dart';
@@ -8,21 +7,21 @@ import '../models/form_field_controller.dart';
 ///
 /// This widget wraps a standard [Form] and connects it to a [FormFieldsController].
 /// It also provides a [FormFieldsController] to its children via [HookedFormProvider].
-class HookedForm<F extends FieldSchema> extends StatelessWidget {
-  /// Creates a [HookedForm] that gets the form from context.
-  const HookedForm({super.key, required this.form, required this.child});
+// ignore: public_member_api_docs
+class const HookedForm<E extends Enum>({
+  super.key,
 
   /// The form controller.
-  final FormFieldsController<F> form;
+  required final FormFieldsController<E> form,
 
   /// The child of the form.
-  final Widget child;
-
+  required final Widget child,
+}) extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Form(
       key: form.key,
-      child: HookedFormProvider<F>(form: form, child: child),
+      child: HookedFormProvider<E>(form: form, child: child),
     );
   }
 }