+++ test/fixtures/test_form_schema.dart
@@ -1,17 +1,4 @@
-
-import 'package:flutter_hook_form/src/models/field_schema.dart';
-import 'package:flutter_hook_form/src/models/validator.dart';
-import 'package:flutter_hook_form/src/validators/validators.dart';
-
-enum TestFormSchema<T> implements FieldSchema<T> {
-  email<String>(validators: [RequiredValidator(), EmailValidator()]),
-  password<String>(validators: [RequiredValidator()]);
-
-  const TestFormSchema({this.validators, this.initialValue});
-
-
-  final T? initialValue;
-
-  @override
-  final List<Validator<T>>? validators;
+enum TestFormSchema {
+  email,
+  password,
 }